Added userID to all database objects and calls

+++ src/dal/transaction/deleteSanitizingDal.ts
@@ -1,5 +1,5 @@
 import { sanitizationModel } from "../../db";
 
-export const deleteSanitizingDal = async (id: string) => {
-  await sanitizationModel.deleteOne({ _id: id });
+export const deleteSanitizingDal = async (id: string, userID: string) => {
+  await sanitizationModel.deleteOne({ _id: id, userID: userID });
 };
